refactor: streamline workload e2e tests for better maintainability
- Rename package from 'replication_test' to 'workload_test' for better clarity
- Reduce test count from 13 to 6 focused tests (54% reduction, -449 lines)
- Remove complex tests (nodeAffinity, stress, mixed workload, priority classes)
- Remove basic pod scheduling test (belongs in basic_scheduling suite)
- Merge resource allocation and queue assignment checks into main ReplicaSet test
- Rename test functions for consistency:
  * Verify_Pod_Failure_Recovery → Verify_ReplicaSet_PodRestart
  * Verify_Deployment_Scaling_Operations → Verify_Deployment_Scheduling
  * Verify_ReplicaSet_Pod_Replication → Verify_ReplicaSet_Scheduling
  * Verify_Rolling_Update_Replication → Verify_Deployment_RollingUpdate

- Move workload-specific helper functions from k8s_utils.go to test file:
  * GetBasicDeploymentSpec, GetBasicStatefulSetSpec, GetBasicHeadlessService
  * Reduces k8s_utils.go bloat and keeps helpers close to usage

- Clean up redundant code:
  * Remove unnecessary 'if len() > 0' checks where length already verified
  * Fix indentation and remove excessive blank lines
  * Add robust cleanup error handling

- Update suite configuration:
  * TestReplication → TestWorkload
  * TEST-replication_junit.xml → TEST-workload_junit.xml
  * Describe('Replication') → Describe('Workload Test')

All 6 remaining tests pass successfully (118s execution time):
✓ Verify_ReplicaSet_Scheduling - ReplicaSet with resource & queue verification
✓ Verify_Deployment_Scheduling - Deployment creation and scaling
✓ Verify_ReplicaSet_PodRestart - Pod failure recovery
✓ Verify_Multi_Container_Pod_Replication - Multi-container pods with init containers
✓ Verify_Deployment_RollingUpdate - Rolling update functionality
✓ Verify_StatefulSet_Replication - StatefulSet with persistent identity

Result: Focused, maintainable test suite with comprehensive workload coverage
